test: improve `proposer-elfs` tests
