+++ crates/proposer-elfs/Cargo.toml
@@ -14,6 +14,6 @@ op-succinct-elfs.workspace = true
 prover-elf-utils.workspace = true
 
 [dev-dependencies]
-alloy-primitives.workspace = true
+insta.workspace = true
 rstest.workspace = true
 sp1-sdk.workspace = true
